Use bodyStream to get raw request body.
Looks like zope publisher provides raw body data via `bodyStream`
attribute of publisher request objects. The documentation for this
feature is not overwhelming and I could not find any guarantees that
this attribute will be available also in future.

It seems, however, to work better and more reliable (in Zope contexts)
than the 'wsgi.input' stream provided in requests.