Provide a test layer for zopish projects.
Layers are a special construct of `zope.testrunner`. Their use is
debatable but we need one here before we are currently using
`zope.testrunner`. Before dropping `zope.testrunner` (and replacing it
with `py.test` or `nose`) we have to provide replacements for the
`BrowserLayer` that works with `zope.testrunner` only.